most of the way towards spawning workers from scripts
.. adding the ComponentLibrary, ComponentFactory and ComponentPair
these classes completely rework how components are created and passed
around.. scripts can pass JSON to the component factory so that it
can get the config params for create.. components created in native
code can just pass the JSON as a string..

.. makes me think that, perhaps, we could use the configurator as the
entire bootstrapping environment.. but not sure how much of the startup
logic i want to expose as a plaintext script that has to be shipped
as separate.. kind of a weird line in the data/logic thing. hm.

.. finally pulled screenWidth/Height/colorDepth out of sugsConfig. whew.

.. made Worker.addComponent() protected, so we are forcing users into the
ComponentFactory scheme.. it makes creation-time configuration of Components
more uniform and allows us to properly inject dependencies by allowing us
to load requirements into the ctor and let the factory sort it out (by creating
needed types or pulling data fromt he configJson)