complete removal of global.sugsConfig . much joy was had.
- in js code, it's completely replaced by worker.current (with one exception
in sugs.coffee where we use sugs.api.core.getPaths() because of a chicken/egg
issue that will be resolved once i have a pure-native embedObjectInNamespace
impl
- in native code, no more pullPropertyFromSugsConfigInGlobal() , which was
only used, really, to pull the worker pointer (set as the private prop on
sugsConfig) for the running environment and do things. Instead, we now set
the worker pointer on the context via JS_SetContextPrivate(), so any native
js func can get at the Worker* .. that solves that.