no more $.startup, rework component setup/teardown
the "startup" should just be done in the top level of the script
that is evaluated when the script is first ran. it's sad that it
took me so much work to make it simpler.

next up: making $.mainLoop not be called from native code and
move to a "true" event loop, of sorts. make doWork() go away
from components and base Worker lifecycle entirely on pending
action in the event loop..