embedObjectInNamespace now has a pure-native impl
instead of native->interpretted .. this caused sugs.coffee to have to
be loaded *before* core functions could be registered (except i *had* to
hand-roll a global getPaths() before-hand, which was lame..) .. this should
allow me to register all of the native bindings before loading sugs.coffee
(which i want to rename to core.coffee, as well), which would be nice.

ultimately, i want core.coffee to just include the definition for require,
would it would be cool if i could ultimately make that a module local value
or somesuch.. but that level of js-fu still escapes me (beyond string cat'ing
together a script, but those is still eluding me, as well.. i think it has to
do with unexpected/early null terminators in combined strings. not sure).