beginning rework of pubsub module
consolidating code paths around whether we are calling into
native code as being a single-target publish or a broadcast
publish, as before.. the diff between fast/durable will just
be a detail that is passed (via the isUdp flag) into the native
func. but we want different native code paths for single-target
for multi-target publishes.